Move builtinInterfaces to BaseContikiMoteType
There is a single caller to getInterfaceClass,
so inline that method at the caller.

This moves the map of builtin interfaces from
MoteInterfaceHandlder to BaseContikiMoteType
where it belongs.
